requirements.txt
pydub
pyupload
pystray
pillow
pyperclip